ipo_predictor/my_flask_app/requirements.txt
Flask
flask-restful
yfinance
requests
python-dotenv
